+++ src/mkvtoolnix-gui/util/settings.h
@@ -30,6 +30,7 @@ public:
     ShutDownComputer,
     HibernateComputer,
     SleepComputer,
+    DeleteSourceFiles,
     Max,
     Default = ExecuteProgram,
   };